+++ src/robot.impl.hh
@@ -152,7 +152,7 @@ namespace hpp
 	virtual void setDimensionExtraConfigSpace (ULong dimension)
 	  throw (hpp::Error);
 
-  virtual Short getDimensionExtraConfigSpace ()
+  virtual ULong getDimensionExtraConfigSpace ()
     throw (hpp::Error);
 
 	virtual void setExtraConfigSpaceBounds
